[FIX] 일지 작성 화면에서 선택한 식물 초기값 설정

+++ feature/diary/src/main/java/com/sabo/feature/diary/plantadd/navigation/PlantAddNavigation.kt
@@ -29,7 +29,7 @@ fun NavController.navigateToPlantEdit(route: PlantAddEdit.PlantEdit) {
 fun NavGraphBuilder.plantAddNavGraph(
     onClickCategory: (String) -> Unit,
     onClickHome: () -> Unit,
-    onClickDiary: () -> Unit,
+    onClickDiary: (Long) -> Unit,
     onClickBack: () -> Unit,
     navigateToHome: () -> Unit
 ) {
